+++ GestionClinique/src/main/java/com/example/GestionClinique/repository/DossierMedicalRepository.java
@@ -4,7 +4,7 @@ import com.example.GestionClinique.model.entity.DossierMedical;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface DossierMedicalRepository extends JpaRepository<DossierMedical, Integer> {
+public interface DossierMedicalRepository extends JpaRepository<DossierMedical, Long> {
