+++ GestionClinique/src/main/java/com/example/GestionClinique/repository/FactureRepository.java
@@ -9,7 +9,7 @@ import java.util.Collection;
 import java.util.List;
 
 
-public interface FactureRepository extends JpaRepository<Facture, Integer> {
+public interface FactureRepository extends JpaRepository<Facture, Long> {
 
 
     Collection<Facture> findFacturesByStatutPaiement(StatutPaiement statutPaiement);
