+++ GestionClinique/src/main/java/com/example/GestionClinique/service/UtilisateurService.java
@@ -9,17 +9,17 @@ import java.util.List;
 public interface UtilisateurService {
 //    Optional<Utilisateur> login(String email, String password);
     Utilisateur createUtilisateur(Utilisateur utilisateur);
-    Utilisateur findUtilisateurById(Integer id);
+    Utilisateur findUtilisateurById(Long id);
     List<Utilisateur> findAllUtilisateur();
-    Utilisateur updateUtilisateur(Integer id, Utilisateur utilisateur);
-    void deleteUtilisateur(Integer id);
+    Utilisateur updateUtilisateur(Long id, Utilisateur utilisateur);
+    void deleteUtilisateur(Long id);
 
-    // Nouvelle méthode: Trouver un utilisateur par son email (généralement unique)
+
     Utilisateur findUtilisateurByEmail(String email);
-    // Nouvelle méthode: Trouver un utilisateur par son nom (peut retourner plusieurs)
+    
     List<Utilisateur> findUtilisateurByNom(String nom);
-    // Nouvelle méthode: Trouver les utilisateurs par rôle
+
     List<Utilisateur> findUtilisateurByRole_RoleType(RoleType roleType);
-    // Nouvelle méthode: Mettre à jour le statut actif/inactif d'un utilisateur
-    Utilisateur updateUtilisateurStatus(Integer id, boolean isActive);
+
+    Utilisateur updateUtilisateurStatus(Long id, boolean isActive);
 }
